refactor: migrate baseline and cache into drift-engine
(cherry picked from commit 1e58d3495243024eab8c6ab92bf9a673958a1106)
